+++ trunk/src/app/srs_app_config.hpp
@@ -481,8 +481,8 @@ private:
 public:
     SrsConfDirective* get_rtc(std::string vhost);
     bool get_rtc_enabled(std::string vhost);
-    bool get_rtc_bframe_discard(std::string vhost);
-    bool get_rtc_aac_discard(std::string vhost);
+    bool get_rtc_keep_bframe(std::string vhost);
+    bool get_rtc_from_rtmp(std::string vhost);
     srs_utime_t get_rtc_stun_timeout(std::string vhost);
     bool get_rtc_stun_strict_check(std::string vhost);
     std::string get_rtc_dtls_role(std::string vhost);